samples/twitter/requirements.txt
textblob
tweepy
pandas
python-dotenv